+++ index.html
@@ -22,18 +22,7 @@
     <meta http-equiv="Content-Type" content="text/html;charset=UTF-8" />
     <link rel="shortcut icon" href="favicon.ico" />
     <link rel="stylesheet" type="text/css" href="main.css">
-	<script src="vector.js" type="text/javascript"></script>
-    <script src="fringe.js" type="text/javascript"></script>
-    <script src="mediaManager.js" type="text/javascript"></script>
-    <script type="text/javascript">
-        window.onload = Initialize;
-
-        function Initialize() {
-            document.getElementById('fringeCanvas').width = window.innerWidth;
-            document.getElementById('fringeCanvas').height = window.innerHeight;
-            Game = new LunaticFringe('fringeCanvas');
-        }
-    </script>
+    <script src="src/initializeGame.js" type="module"></script>
 </head>
 <body>
     <canvas id="fringeCanvas" style="position: absolute; top: 0px; left: 0px;"></canvas>
